linera_sdk

Module abi

This module defines the notion of Application Binary Interface (ABI) for Linera applications across Wasm and native architectures.

Traits

Abi

A trait that includes all the types exported by a Linera application (both contract and service).

ContractAbi

A trait that includes all the types exported by a Linera application contract.

ServiceAbi

A trait that includes all the types exported by a Linera application service.

WithContractAbi

Marker trait to help importing contract types.

WithServiceAbi

Marker trait to help importing service types.