Module linera_sdk::base

Types reexported from linera_base.

Structs

Account

A system account.

Amount

A non-negative amount of tokens.

ApplicationId

A unique identifier for a user application.

ApplicationPermissions

Permissions for applications on a chain.

BlockHeight

A block height to identify blocks in a chain.

BytecodeId

A unique identifier for an application bytecode.

ChainId

The unique identifier (UID) of a chain. This is currently computed as the hash value of a ChainDescription.

ChannelName

The name of a subscription channel.

MessageId

The index of a message in a chain.

Owner

The owner of a chain. This is currently the hash of the owner’s public key used to verify signatures.

Resources

Resources that an application may spend during the execution of transaction or an application call.

SendMessageRequest

A request to send a message.

Timestamp

A timestamp, in microseconds since the Unix epoch.

Enums

AccountOwner

An account owner.

ArithmeticError

An error type for arithmetic errors.

BcsHexParseError

An error trying to parse the hex-digits of a BCS-encoded value.

ChainDescription

How to create a chain.

Destination

The destination of a message, relative to a particular application.

GenericApplicationId

A unique identifier for an application.

ParseAmountError

Round

An identifier for successive attempts to decide a value in a consensus protocol.

Traits

Abi

A trait that includes all the types exported by a Linera application (both contract and service).

ContractAbi

A trait that includes all the types exported by a Linera application contract.

ServiceAbi

A trait that includes all the types exported by a Linera application service.

WithContractAbi

Marker trait to help importing contract types.

WithServiceAbi

Marker trait to help importing service types.