Trait linera_sdk::base::BcsSignable

pub trait BcsSignable<'de>: Serialize + Deserialize<'de> { }

Activate the blanket implementation of Signable based on serde and BCS.

We use serde_name to extract a seed from the name of structs and enums.

We use BCS to generate canonical bytes suitable for signing.

Object Safety

This trait is not object safe.