Trait linera_sdk::base::ContractAbi

pub trait ContractAbi {
    type InitializationArgument: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type Parameters: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type Operation: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type Effect: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type ApplicationCall: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type SessionCall: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type SessionState: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type Response: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
}

A trait that includes all the types exported by a Linera application contract.

Required Associated Types

type InitializationArgument: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

Initialization argument passed to a new application on the chain that created it (e.g. an initial amount of tokens minted).

To share configuration data on every chain, use ContractAbi::Parameters instead.

type Parameters: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

Immutable parameters specific to this application (e.g. the name of a token).

type Operation: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type of operation executed by the application.

Operations are transactions directly added to a block by the creator (and signer) of the block. Users typically use operations to start interacting with an application on their own chain.

type Effect: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type of effect executed by the application.

Effects are executed when a message created by the same application is received from another chain and accepted in a block.

type ApplicationCall: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The argument type when this application is called from another application on the same chain.

type SessionCall: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The argument type when a session of this application is called from another application on the same chain.

Sessions are temporary objects that may be spawned by an application call. Once created, they must be consumed before the current transaction ends.

type SessionState: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type for the state of a session.

type Response: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The response type of an application call.