Module linera_sdk::contract::exported_futures

Types for the exported futures for the contract endpoints.

Each type is called by the code generated by [wit-bindgen-guest-rust] when the host calls the guest WASM module’s respective endpoint. This module contains the code to forward the call to the contract type that implements Contract.

Structs

ExecuteEffect

Future implementation exported from the guest to allow the host to call Contract::execute_effect.

ExecuteOperation

Future implementation exported from the guest to allow the host to call Contract::execute_operation.

HandleApplicationCall

Future implementation exported from the guest to allow the host to call Contract::handle_application_call.

HandleSessionCall

Future implementation exported from the guest to allow the host to call Contract::handle_session_call.

Initialize

Future implementation exported from the guest to allow the host to call Contract::initialize.

Traits

ContractStateStorage

The storage APIs used by a contract.