Trait linera_sdk::contract::exported_futures::ContractStateStorage

pub trait ContractStateStorage<Application> {
    fn load_and_lock<'async_trait>(
    ) -> Pin<Box<dyn Future<Output = Application> + Send + 'async_trait>>;
    fn store_and_unlock<'async_trait>(
        state: Application
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>;
    fn execute_with_released_state<'life0, 'async_trait, Operation>(
        state: &'life0 mut Application,
        operation: impl 'async_trait + FnOnce() -> Operation + Send
    ) -> Pin<Box<dyn Future<Output = Operation::Output> + Send + 'async_trait>>
    where
        Operation: Future + Send,
        Operation::Output: Send,
        Operation: 'async_trait,
        'life0: 'async_trait;

    fn execute_with_state<'async_trait, Operation, Success, Error>(
        operation: Operation
    ) -> Pin<Box<dyn Future<Output = Result<Success, String>> + Send + 'async_trait>>
    where
        Operation: for<'app> FnOnce(&'app mut Application) -> Pin<Box<dyn Future<Output = Result<Success, Error>> + Send + 'app>>,
        Application: Send,
        Operation: Send,
        Success: Send + 'static,
        Error: ToString + 'static,
        Operation: 'async_trait,
        Success: 'async_trait,
        Error: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

The storage APIs used by a contract.

Required Methods

fn load_and_lock<'async_trait>(
) -> Pin<Box<dyn Future<Output = Application> + Send + 'async_trait>>

Loads the Application state and locks it for writing.

fn store_and_unlock<'async_trait>(
state: Application
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>

Stores the Application state and unlocks it for reads and writes.

fn execute_with_released_state<'life0, 'async_trait, Operation>(
state: &'life0 mut Application,
operation: impl 'async_trait + FnOnce() -> Operation + Send
) -> Pin<Box<dyn Future<Output = Operation::Output> + Send + 'async_trait>>where
Operation: Future + Send,
Operation::Output: Send,
Operation: 'async_trait,
'life0: 'async_trait,

Executes an operation, persisting the Application state before execution and reloading the state afterwards.

Provided Methods

fn execute_with_state<'async_trait, Operation, Success, Error>(
operation: Operation
) -> Pin<Box<dyn Future<Output = Result<Success, String>> + Send + 'async_trait>>where
Operation: for<'app> FnOnce(&'app mut Application) -> Pin<Box<dyn Future<Output = Result<Success, Error>> + Send + 'app>>,
Application: Send,
Operation: Send,
Success: Send + 'static,
Error: ToString + 'static,
Operation: 'async_trait,
Success: 'async_trait,
Error: 'async_trait,
Self: Send + 'async_trait,

Executes an operation with the Application state.

The state is only stored back in storage if the operation succeeds. Otherwise, the error is returned as a String.