Trait linera_sdk::contract::Contract

pub trait Contract {
    type Initialize: Initialize;
    type ExecuteOperation: ExecuteOperation;
    type ExecuteEffect: ExecuteEffect;
    type HandleApplicationCall: HandleApplicationCall;
    type HandleSessionCall: HandleSessionCall;

    fn drop_initialize(val: Self::Initialize) { ... }
    fn drop_execute_operation(val: Self::ExecuteOperation) { ... }
    fn drop_execute_effect(val: Self::ExecuteEffect) { ... }
    fn drop_handle_application_call(val: Self::HandleApplicationCall) { ... }
    fn drop_handle_session_call(val: Self::HandleSessionCall) { ... }
}

Required Associated Types

type Initialize: Initialize

type ExecuteOperation: ExecuteOperation

type ExecuteEffect: ExecuteEffect

type HandleApplicationCall: HandleApplicationCall

type HandleSessionCall: HandleSessionCall

Provided Methods

fn drop_initialize(val: Self::Initialize)

An optional callback invoked when a handle is finalized and destroyed.

fn drop_execute_operation(val: Self::ExecuteOperation)

An optional callback invoked when a handle is finalized and destroyed.

fn drop_execute_effect(val: Self::ExecuteEffect)

An optional callback invoked when a handle is finalized and destroyed.

fn drop_handle_application_call(val: Self::HandleApplicationCall)

An optional callback invoked when a handle is finalized and destroyed.

fn drop_handle_session_call(val: Self::HandleSessionCall)

An optional callback invoked when a handle is finalized and destroyed.