Module linera_sdk::ethereum

Support for Linera applications that interact with Ethereum or other EVM contracts.

Structs

EthereumClient

A wrapper for a URL that implements JsonRpcClient and uses the JSON oracle to make requests.

EthereumEvent

The data type for an Ethereum event emitted by a smart contract

Enums

EthereumDataType

A single primitive data type. This is used for example for the entries of Ethereum events.

Traits

EthereumQueries

The basic Ethereum queries that can be used from a smart contract and do not require gas to be executed.