linera_sdk::linera_base_types::signer

Trait Signer

pub trait Signer: Send {
    type Error: Error;

    // Required methods
    fn sign(
        &self,
        owner: &AccountOwner,
        value: &CryptoHash,
    ) -> impl Future<Output = Result<AccountSignature, Self::Error>> + Send;
    fn contains_key(
        &self,
        owner: &AccountOwner,
    ) -> impl Future<Output = Result<bool, Self::Error>> + Send;
}

A trait for signing keys.

Required Associated Types

type Error: Error

The type of errors arising from operations on this Signer.

Required Methods

fn sign( &self, owner: &AccountOwner, value: &CryptoHash, ) -> impl Future<Output = Result<AccountSignature, Self::Error>> + Send

Creates a signature for the given value using the provided owner.

fn contains_key( &self, owner: &AccountOwner, ) -> impl Future<Output = Result<bool, Self::Error>> + Send

Returns whether the given owner is a known signer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.