linera_sdk::linera_base_types

Trait BcsSignable

pub trait BcsSignable<'de>: Serialize + Deserialize<'de> { }

Activate the blanket implementation of Signable based on serde and BCS.

We use serde_name to extract a seed from the name of structs and enums.

We use BCS to generate canonical bytes suitable for signing.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.