linera_sdk::linera_base_types
Trait Signer

pub trait Signer: Send + Sync {
    // Required methods
    fn sign<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        owner: &'life1 AccountOwner,
        value: &'life2 CryptoHash,
    ) -> Pin<Box<dyn Future<Output = Result<AccountSignature, Box<dyn Error>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_public_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        owner: &'life1 AccountOwner,
    ) -> Pin<Box<dyn Future<Output = Result<AccountPublicKey, Box<dyn Error>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn contains_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        owner: &'life1 AccountOwner,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Box<dyn Error>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
}

A trait for signing keys.

Required Methods

fn sign<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, owner: &'life1 AccountOwner, value: &'life2 CryptoHash, ) -> Pin<Box<dyn Future<Output = Result<AccountSignature, Box<dyn Error>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Creates a signature for the given value using the provided owner.

fn get_public_key<'life0, 'life1, 'async_trait>( &'life0 self, owner: &'life1 AccountOwner, ) -> Pin<Box<dyn Future<Output = Result<AccountPublicKey, Box<dyn Error>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Returns the public key corresponding to the given owner.

fn contains_key<'life0, 'life1, 'async_trait>( &'life0 self, owner: &'life1 AccountOwner, ) -> Pin<Box<dyn Future<Output = Result<bool, Box<dyn Error>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Returns whether the given owner is a known signer.