Module linera_sdk::service

Types and macros useful for writing an application service.

Modules

exported_futures

Types for the exported futures for the service endpoints.

system_api

Functions and types to interface with the system API available to application services.

wit_types

Internal module with code generated by wit-bindgen.