Module linera_sdk::service::system_api

Functions and types to interface with the system API available to application contracts.

Structs

ReadOnlyKeyValueStore

A type to interface in a read-only manner with the key value storage provided to applications.

Functions

current_application_id

Retrieve the current application ID.

current_application_parameters

Retrieve the current application parameters.

current_chain_id

Retrieve the current chain ID.

current_system_balance

Retrieve the current system balance.

current_system_time

Retrieves the current system time.

load

Load the contract state, without locking it for writes.

load_view_using

Helper function to load the service state or create a new one if it doesn’t exist.

lock_and_load_view

Load the service state, without locking it for writes.

log

Requests the host to log a message.

query_application

Queries another application.

unlock_view

Load the service state, without locking it for writes.

Type Definitions

ReadOnlyViewStorageContext

Implementation of linera_views::common::Context that uses the ReadOnlyKeyValueStore to allow views to read data from the storage layer provided to Linera applications.