Trait linera_sdk::Contract

pub trait Contract: WithContractAbi + ContractAbi + Send + Sized {
    type Error: Error + From<Error> + From<Error> + 'static;
    type State: Sync;
    type Storage: ContractStateStorage<Self> + Send + 'static;
    type Message: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;

    // Required methods
    fn new<'async_trait>(
        state: Self::State,
        runtime: ContractRuntime<Self>
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn state_mut(&mut self) -> &mut Self::State;
    fn initialize<'life0, 'async_trait>(
        &'life0 mut self,
        argument: Self::InitializationArgument
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute_operation<'life0, 'async_trait>(
        &'life0 mut self,
        operation: Self::Operation
    ) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute_message<'life0, 'async_trait>(
        &'life0 mut self,
        message: Self::Message
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn finalize<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

The contract interface of a Linera application.

As opposed to the Service interface of an application, contract entry points are triggered by the execution of blocks in a chain. Their execution may modify storage and is gas-metered.

Below we use the word “transaction” to refer to the current operation or message being executed.

Required Associated Types

type Error: Error + From<Error> + From<Error> + 'static

The type used to report errors to the execution environment.

Errors are not recoverable and always interrupt the current transaction. To return recoverable errors in the case of application calls, you may use the response types.

type State: Sync

The type used to store the persisted application state.

type Storage: ContractStateStorage<Self> + Send + 'static

The desired storage backend used to store the application’s state.

Currently, the two supported backends are SimpleStateStorage or ViewStateStorage. Accordingly, this associated type may be defined as type Storage = SimpleStateStorage<Self> or type Storage = ViewStateStorage<Self>.

The first deployment on other chains will use the Default implementation of the application state if SimpleStateStorage is used, or the Default value of all sub-views in the state if the ViewStateStorage is used.

type Message: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type of message executed by the application.

Messages are executed when a message created by the same application is received from another chain and accepted in a block.

Required Methods

fn new<'async_trait>( state: Self::State, runtime: ContractRuntime<Self> ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait,

Creates a in-memory instance of the contract handler from the application’s state.

fn state_mut(&mut self) -> &mut Self::State

Returns the current state of the application so that it can be persisted.

fn initialize<'life0, 'async_trait>( &'life0 mut self, argument: Self::InitializationArgument ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Initializes the application on the chain that created it.

This is only called once when the application is created and only on the microchain that created the application.

fn execute_operation<'life0, 'async_trait>( &'life0 mut self, operation: Self::Operation ) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Applies an operation from the current block.

Operations are created by users and added to blocks, serving as the starting point for an application’s execution.

fn execute_message<'life0, 'async_trait>( &'life0 mut self, message: Self::Message ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Applies a message originating from a cross-chain message.

Messages are messages sent across chains. These messages are created and received by the same application. Messages can be either single-sender and single-receiver, or single-sender and multiple-receivers. The former allows sending cross-chain messages to the application on some other specific chain, while the latter uses broadcast channels to send a message to multiple other chains where the application is subscribed to a sender channel on this chain.

For a message to be executed, a user must mark it to be received in a block of the receiver chain.

Provided Methods

fn finalize<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Finishes the execution of the current transaction.

This is called once before a transaction ends, to allow all applications that participated in the transaction to perform any final operations, and optionally it may also cancel the transaction if there are any pendencies.

The default implementation persists the state, so if this method is overriden, care must be taken to persist the state manually.

Object Safety

This trait is not object safe.