Trait linera_sdk::Service

pub trait Service: WithServiceAbi + ServiceAbi + Sized {
    type Error: Error + From<Error>;
    type State;
    type Storage: ServiceStateStorage;

    // Required methods
    async fn new(
        state: Self::State,
        runtime: ServiceRuntime<Self>
    ) -> Result<Self, Self::Error>;
    async fn handle_query(
        &self,
        query: Self::Query
    ) -> Result<Self::QueryResponse, Self::Error>;
}

The service interface of a Linera application.

As opposed to the Contract interface of an application, service entry points are triggered by JSON queries (typically GraphQL). Their execution cannot modify storage and is not gas-metered.

Required Associated Types

type Error: Error + From<Error>

Type used to report errors to the execution environment.

Errors are not recoverable and always interrupt the current query.

type State

The type used to store the persisted application state.

type Storage: ServiceStateStorage

The desired storage backend used to store the application’s state.

Currently, the two supported backends are SimpleStateStorage or ViewStateStorage. Accordingly, this associated type may be defined as type Storage = SimpleStateStorage<Self> or type Storage = ViewStateStorage<Self>.

Required Methods

async fn new( state: Self::State, runtime: ServiceRuntime<Self> ) -> Result<Self, Self::Error>

Creates a in-memory instance of the service handler from the application’s state.

async fn handle_query( &self, query: Self::Query ) -> Result<Self::QueryResponse, Self::Error>

Executes a read-only query on the state of this application.

Object Safety

This trait is not object safe.