linera_sdk

Trait ToBcsBytes

pub trait ToBcsBytes {
    // Required method
    fn to_bcs_bytes(&self) -> Result<Vec<u8>, Error>;
}

Extension trait to serialize a type into a vector of bytes using bcs.

Required Methods

fn to_bcs_bytes(&self) -> Result<Vec<u8>, Error>

Serializes itself into a vector of bytes using bcs.