Module linera_sdk::views

Helper types for using linera_views to store application state.

Structs

KeyValueStore

A type to interface with the key value storage provided to applications.

Type Definitions

ViewStorageContext

Implementation of linera_views::common::Context that uses the KeyValueStore to allow views to store data in the storage layer provided to Linera applications.