Struct linera_sdk::views::batch::Batch

pub struct Batch {
    pub operations: Vec<WriteOperation, Global>,
}

A batch of writes inside a transaction.

Fields
operations: Vec<WriteOperation, Global>

The entries of the batch to be consumed when processed.

Implementations

impl Batch

pub fn new() -> Batch

Creates an empty batch.

pub async fn build<F>( builder: F ) -> impl Future<Output = Result<Batch, ViewError>>where F: FnOnce(&mut Batch) -> Pin<Box<dyn Future<Output = Result<(), ViewError>> + Send, Global>> + Send + Sync,

Builds a batch from a builder function.

pub fn simplify(self) -> UnorderedBatch

Simplifies the batch by removing operations that are overwritten by others.

A key may appear multiple times in the batch, as an insert, a delete or matched by a delete prefix.

  let mut batch = Batch::new();
  batch.put_key_value(vec![0,1], &(34 as u128));
  batch.delete_key(vec![0,1]);
  let unord_batch = batch.simplify();
  assert_eq!(unord_batch.key_prefix_deletions.len(), 0);
  assert_eq!(unord_batch.simple_unordered_batch.insertions.len(), 0);
  assert_eq!(unord_batch.simple_unordered_batch.deletions.len(), 1);

pub fn put_key_value( &mut self, key: Vec<u8, Global>, value: &impl Serialize ) -> Result<(), Error>

Inserts the insertion of a (key,value) pair into the batch with a serializable value.

  let mut batch = Batch::new();
  batch.put_key_value(vec![0,1], &(34 as u128));

pub fn put_key_value_bytes( &mut self, key: Vec<u8, Global>, value: Vec<u8, Global> )

Inserts the insertion of a (key,value) pair into the batch with value a vector of u8.

  let mut batch = Batch::new();
  batch.put_key_value_bytes(vec![0,1], vec![3,4,5]);

pub fn delete_key(&mut self, key: Vec<u8, Global>)

Inserts the deletion of a key into the batch.

  let mut batch = Batch::new();
  batch.delete_key(vec![0,1]);

pub fn delete_key_prefix(&mut self, key_prefix: Vec<u8, Global>)

Inserts the deletion of a key_prefix into the batch.

  let mut batch = Batch::new();
  batch.delete_key_prefix(vec![0,1]);