Struct linera_sdk::views::collection_view::CollectionView

pub struct CollectionView<C, I, W> { /* private fields */ }

A view that supports accessing a collection of views of the same kind, indexed by a key, one subview at a time.

Implementations

impl<C, I, W> CollectionView<C, I, W>where C: Context + Send, ViewError: From<<C as Context>::Error>, I: Serialize, W: View<C>,

pub async fn load_entry_mut<Q>( &mut self, index: &Q ) -> impl Future<Output = Result<&mut W, ViewError>>where I: Borrow<Q>, Q: Serialize + ?Sized,

Loads a subview for the data at the given index in the collection. If an entry was removed before then a default entry is put on this index.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn load_entry<Q>( &mut self, index: &Q ) -> impl Future<Output = Result<&W, ViewError>>where I: Borrow<Q>, Q: Serialize + ?Sized,

Loads a subview for the data at the given index in the collection. If an entry was removed before then a default entry is put on this index.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  let subview = view.load_entry(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn try_load_entry<Q>( &self, index: &Q ) -> impl Future<Output = Result<ReadGuardedView<'_, W>, ViewError>>where I: Borrow<Q>, Q: Serialize + ?Sized,

Same as load_entry_mut but for read-only access. May fail if one subview is already being visited.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  let subview = view.try_load_entry(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn reset_entry_to_default<Q>( &mut self, index: &Q ) -> impl Future<Output = Result<(), ViewError>>where I: Borrow<Q>, Q: Serialize + ?Sized,

Resets an entry to the default value.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  *value = String::from("Hello");
  view.reset_entry_to_default(&23).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  assert_eq!(*value, String::default());

pub fn remove_entry<Q>(&mut self, index: &Q) -> Result<(), ViewError>where I: Borrow<Q>, Q: Serialize + ?Sized,

Removes an entry from the CollectionView. If absent nothing happens.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  assert_eq!(*value, String::default());
  view.remove_entry(&23);
  let keys = view.indices().await.unwrap();
  assert_eq!(keys.len(), 0);

pub fn extra(&self) -> &<C as Context>::Extra

Gets the extra data.

impl<C, I, W> CollectionView<C, I, W>where C: Context + Send, ViewError: From<<C as Context>::Error>, I: Sync + Clone + Send + Debug + Serialize + DeserializeOwned, W: View<C> + Sync,

pub async fn indices( &self ) -> impl Future<Output = Result<Vec<I, Global>, ViewError>>

Returns the list of indices in the collection in the order determined by the serialization.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  view.load_entry_mut(&25).await.unwrap();
  let indices = view.indices().await.unwrap();
  assert_eq!(indices.len(), 2);

impl<C, I, W> CollectionView<C, I, W>where C: Context + Send, ViewError: From<<C as Context>::Error>, I: Debug + DeserializeOwned, W: View<C> + Sync,

pub async fn for_each_index_while<F>( &self, f: F ) -> impl Future<Output = Result<(), ViewError>>where F: FnMut(I) -> Result<bool, ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the serialization. If the function returns false then the loop ends prematurely.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  view.load_entry_mut(&24).await.unwrap();
  let mut count = 0;
  view.for_each_index_while(|_key| {
    count += 1;
    Ok(count < 1)
  }).await.unwrap();
  assert_eq!(count, 1);

pub async fn for_each_index<F>( &self, f: F ) -> impl Future<Output = Result<(), ViewError>>where F: FnMut(I) -> Result<(), ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the serialization.

  let mut view : CollectionView<_, u64, RegisterView<_,String>> = CollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  view.load_entry_mut(&28).await.unwrap();
  let mut count = 0;
  view.for_each_index(|_key| {
    count += 1;
    Ok(())
  }).await.unwrap();
  assert_eq!(count, 2);