Trait linera_sdk::views::crypto::BcsSignable

pub trait BcsSignable: Serialize + DeserializeOwned { }

Activate the blanket implementation of Signable based on serde and BCS.

We use serde_name to extract a seed from the name of structs and enums.

We use BCS to generate canonical bytes suitable for signing.