Struct linera_sdk::views::futures::io::AllowStdIo

pub struct AllowStdIo<T>(_);

A simple wrapper type which allows types which implement only implement std::io::Read or std::io::Write to be used in contexts which expect an AsyncRead or AsyncWrite.

If these types issue an error with the kind io::ErrorKind::WouldBlock, it is expected that they will notify the current task on readiness. Synchronous std types should not issue errors of this kind and are safe to use in this context. However, using these types with AllowStdIo will cause the event loop to block, so they should be used with care.

Implementations

impl<T> AllowStdIo<T>

pub fn new(io: T) -> AllowStdIo<T>

Creates a new AllowStdIo from an existing IO object.

pub fn get_ref(&self) -> &T

Returns a reference to the contained IO object.

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the contained IO object.

pub fn into_inner(self) -> T

Consumes self and returns the contained IO object.