Module linera_sdk::views::test_utils

Helper types for tests.

Structs

LocalStackTestContext

A type to help tests that need a LocalStack instance.

Functions

get_random_byte_vector

Takes a random number generator, a key_prefix and extends it by n random bytes.

get_random_key_value_operations

A bunch of puts and some deletes.

get_random_key_value_vec

Takes a random number generator rng, a number n and returns n random (key,value) which are all distinct with key and value are of length 8.

get_random_key_value_vec_prefix

Takes a random number generator, a key_prefix and generates pairs (key,value) with key obtained by appending 8 bytes at random to key_prefix and value obtained by appending 8 bytes to the trivial vector. We return n such (key,value) pairs which are all distinct

get_random_kset

Builds a random k element subset of n

list_buckets

Helper function to list the names of buckets registered on S3.

list_tables

Helper function to list the names of tables registered on DynamoDB.

random_shuffle

Shuffles the values entries randomly

span_random_reordering_put_delete

A random reordering of the puts and deletes. For something like MapView it should get us the same result whatever way we are calling.