Trait linera_sdk::views::View

pub trait View<C>: Sized {
    // Required methods
    fn context(&self) -> &C;
    fn load<'async_trait>(
        context: C
    ) -> Pin<Box<dyn Future<Output = Result<Self, ViewError>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn rollback(&mut self);
    fn clear(&mut self);
    fn flush(&mut self, batch: &mut Batch) -> Result<bool, ViewError>;
}

A view gives exclusive access to read and write the data stored at an underlying address in storage.

Required Methods

fn context(&self) -> &C

Obtains a mutable reference to the internal context.

fn load<'async_trait>( context: C ) -> Pin<Box<dyn Future<Output = Result<Self, ViewError>> + Send + 'async_trait>>
where Self: 'async_trait,

Creates a view or a subview.

fn rollback(&mut self)

Discards all pending changes. After that flush should have no effect to storage.

fn clear(&mut self)

Clears the view. That can be seen as resetting to default. If the clear is followed by a flush then all the relevant data is removed on the storage.

fn flush(&mut self, batch: &mut Batch) -> Result<bool, ViewError>

Persists changes to storage. This leaves the view still usable and is essentially neutral to the program running. Crash-resistant storage implementations are expected to accumulate the desired changes in the batch variable first. If the view is dropped without calling flush, staged changes are simply lost. The returned boolean indicates whether the operation removes the view or not.

Object Safety

This trait is not object safe.