Module linera_sdk::views::views

The definition of the View and related traits.

Enums

ViewError

Main error type for the crate.

Traits

CryptoHashRootView

A RootView that also supports crypto hash

CryptoHashView

A View that also supports crypto hash

HashableView

A view that supports hashing its values.

Hasher

The requirement for the hasher type in HashableView.

RootView

A View whose staged modifications can be saved in storage.

View

A view gives exclusive access to read and write the data stored at an underlying address in storage.

Derive Macros

CryptoHashRootView

CryptoHashView

GraphQLView

HashableView

RootView

View