Crate linera_sdk

This module provides an SDK for developing Linera applications using Rust.

A Linera application consists of two WebAssembly binaries: a contract and a service. In both binaries, there should be a shared application state. The state is a type that represents what the application would like to persist in storage across blocks, and must implement State trait in order to specify how the state should be loaded from and stored to the persistent key-value storage. An alternative is to use the linera-views, a framework that allows loading selected parts of the state. This is useful if the application’s state is large and doesn’t need to be loaded in its entirety for every execution. By deriving RootView on the state type it automatically implements the State trait.

The contract binary should create a type to implement the Contract trait. The type can store the ContractRuntime and the state, and must have its implementation exported by using the contract! macro.

The service binary should create a type to implement the Service trait. The type can store the ServiceRuntime and the state, and must have its implementation exported by using the service! macro.

Examples

The examples directory contains some example applications.

Re-exports

pub use self::mock_system_api::MockSystemApi;

pub use self::contract::ContractRuntime;

pub use self::service::ServiceRuntime;

Modules

abi

This module defines the notion of Application Binary Interface (ABI) for Linera applications across Wasm and native architectures.

abis

Common ABIs that may have multiple implementations.

base

Types reexported from linera_base.

contract

Types and macros useful for writing an application contract.

graphql

GraphQL traits for generating interfaces into applications.

mock_system_api

Hooks for mocking system APIs inside unit tests.

service

Types and macros useful for writing an application service.

test

Helper types for writing integration tests for WebAssembly applications.

util

Module with helper types and functions used by the SDK.

views

Helper types for using linera_views to store application state.

Macros

contract

Declares an implementation of the Contract trait, exporting it from the Wasm module.

ensure

A macro for asserting that a condition is true, returning an error if it is not.

service

Declares an implementation of the Service trait, exporting it from the Wasm module.

Structs

ContractLogger

A logger that uses the system API for contracts.

EmptyState

Representation of an empty persistent state.

Resources

Resources that an application may spend during the execution of transaction or an application call.

SendMessageRequest

A request to send a message.

ServiceLogger

A logger that uses the system API for services.

Traits

Contract

The contract interface of a Linera application.

FromBcsBytes

Extension trait to deserialize a type from a vector of bytes using bcs.

Service

The service interface of a Linera application.

State

The persistent state of a Linera application.

ToBcsBytes

Extension trait to serialize a type into a vector of bytes using bcs.