Crate linera_sdk

This module provides an SDK for developing Linera applications using Rust.

A Linera application consists of two WebAssembly binaries: a contract and a service. In both binaries, there should be a shared application state. The state is a type that represents what the application would like to persist in storage across blocks, and must implement the Contract trait in the contract binary and the Service trait in the service binary.

The application can select between two storage backends to use. Selecting the storage backend is done by specifying both the Contract::Storage and the Service::Storage associated types.

The SimpleStateStorage backend stores the application’s state type by serializing it into binary blob. This allows the entire contents of the state to be persisted and made available to the application when it is executed.

The ViewStateStorage backend stores the application’s state using the linera-views, a framework that allows loading selected parts of the state. This is useful if the application’s state is large and doesn’t need to be loaded in its entirety for every execution.

The contract binary should use the contract! macro to export the application’s contract endpoints implemented via the Contract trait implementation.

The service binary should use the service! macro to export the application’s service endpoints implemented via the Service trait implementation.

Examples

The examples directory contains some example applications.

Modules

base

Types reexported from linera_base.

contract

Types and macros useful for writing an application contract.

service

Types and macros useful for writing an application service.

test

Helper types for writing integration tests for WebAssembly applications.

Macros

contract

Declares an implementation of the Contract trait, exporting it from the WASM module.

ensure

A macro for asserting that a condition is true, returning an error if it is not.

export_contract

Declares the export of the interface for the given type.

export_service

Declares the export of the interface for the given type.

instance_exported_future

Declares a type exporting a future declared as a WIT type.

service

Declares an implementation of the Service trait, exporting it from the WASM module.

Structs

ApplicationCallResult

The result of calling into a user application.

CalleeContext

The context of the execution of an application’s cross-application call or session call handler.

ContractLogger

A logger that uses the system API for contracts.

EffectContext

The context of the execution of an application’s effect.

ExecutionResult

Externally visible results of an execution. These results are meant in the context of the application that created them.

ExportedFuture

A future that’s exported from a guest WASM module to the host.

OperationContext

The context of the execution of an application’s operation.

QueryContext

The context of the execution of an application’s query.

ServiceLogger

A logger that uses the system API for services.

Session

Syscall to request creating a new session.

SessionCallResult

The result of calling into a session.

SimpleStateStorage

A simple state management runtime using a single byte array.

ViewStateStorage

A state management runtime based on linera-views.

Traits

Contract

The public entry points provided by an application’s contract.

FromBcsBytes

Extension trait to deserialize a type from a vector of bytes using bcs.

Service

The public entry points provided by an application’s service.

ToBcsBytes

Extension trait to serialize a type into a vector of bytes using bcs.