linera_sdk::abi

Trait ServiceAbi

pub trait ServiceAbi {
    type Query: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type QueryResponse: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
}

A trait that includes all the types exported by a Linera application service.

Required Associated Types

type Query: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type of a query receivable by the application’s service.

type QueryResponse: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The response type of the application’s service.