Module linera_sdk::base

Structs

Amount

ApplicationId

A unique identifier for an application.

BlockHeight

A block height to identify blocks in a chain.

BytecodeId

A unique identifier for an application bytecode.

ChainId

The unique identifier (UID) of a chain. This is the hash value of a ChainDescription.

ChannelName

The name of a subscription channel.

CryptoHash

A Sha3-256 value.

EffectId

The index of an effect in a chain.

Owner

The owner of an account (aka address).

SessionId

The identifier of a session.

SystemBalance

The balance of a chain.

Timestamp

A timestamp, in microseconds since the Unix epoch.

Enums

ArithmeticError

An error type for arithmetic errors.

Destination

The destination of a message, relative to a particular application.

Traits

BcsSignable

Activate the blanket implementation of Signable based on serde and BCS.

Functions

hex_debug

Prints a vector of bytes in hexadecimal.