Trait linera_sdk::base::ContractAbi

pub trait ContractAbi {
    type Operation: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
    type Response: Serialize + DeserializeOwned + Send + Sync + Debug + 'static;
}

A trait that includes all the types exported by a Linera application contract.

Required Associated Types

type Operation: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The type of operation executed by the application.

Operations are transactions directly added to a block by the creator (and signer) of the block. Users typically use operations to start interacting with an application on their own chain.

type Response: Serialize + DeserializeOwned + Send + Sync + Debug + 'static

The response type of an application call.