Struct linera_sdk::contract::MockContractRuntime

pub struct MockContractRuntime<Application>
where
    Application: Contract,
{ /* private fields */ }

A mock of the common runtime to interface with the host executing the contract.

Implementations

impl<Application> MockContractRuntime<Application>
where Application: Contract,

pub fn new() -> Self

Creates a new MockContractRuntime instance for a contract.

pub fn key_value_store(&self) -> KeyValueStore

Returns the key-value store to interface with storage.

pub fn root_view_storage_context(&self) -> ViewStorageContext

Returns a storage context suitable for a root view.

pub fn with_application_parameters( self, application_parameters: Application::Parameters, ) -> Self

Configures the application parameters to return during the test.

pub fn set_application_parameters( &mut self, application_parameters: Application::Parameters, ) -> &mut Self

Configures the application parameters to return during the test.

pub fn application_parameters(&mut self) -> Application::Parameters

Returns the application parameters provided when the application was created.

pub fn with_application_id( self, application_id: ApplicationId<Application::Abi>, ) -> Self

Configures the application ID to return during the test.

pub fn set_application_id( &mut self, application_id: ApplicationId<Application::Abi>, ) -> &mut Self

Configures the application ID to return during the test.

pub fn application_id(&mut self) -> ApplicationId<Application::Abi>

Returns the ID of the current application.

pub fn with_application_creator_chain_id(self, chain_id: ChainId) -> Self

Configures the application creator chain ID to return during the test.

pub fn set_application_creator_chain_id( &mut self, chain_id: ChainId, ) -> &mut Self

Configures the application creator chain ID to return during the test.

pub fn application_creator_chain_id(&mut self) -> ChainId

Returns the chain ID of the current application creator.

pub fn with_chain_id(self, chain_id: ChainId) -> Self

Configures the chain ID to return during the test.

pub fn set_chain_id(&mut self, chain_id: ChainId) -> &mut Self

Configures the chain ID to return during the test.

pub fn chain_id(&mut self) -> ChainId

Returns the ID of the current chain.

pub fn with_authenticated_signer( self, authenticated_signer: impl Into<Option<Owner>>, ) -> Self

Configures the authenticated signer to return during the test.

pub fn set_authenticated_signer( &mut self, authenticated_signer: impl Into<Option<Owner>>, ) -> &mut Self

Configures the authenticated signer to return during the test.

pub fn authenticated_signer(&mut self) -> Option<Owner>

Returns the authenticated signer for this execution, if there is one.

pub fn with_block_height(self, block_height: BlockHeight) -> Self

Configures the block height to return during the test.

pub fn set_block_height(&mut self, block_height: BlockHeight) -> &mut Self

Configures the block height to return during the test.

pub fn block_height(&mut self) -> BlockHeight

Returns the height of the current block that is executing.

pub fn with_message_id(self, message_id: impl Into<Option<MessageId>>) -> Self

Configures the message ID to return during the test.

pub fn set_message_id( &mut self, message_id: impl Into<Option<MessageId>>, ) -> &mut Self

Configures the message ID to return during the test.

pub fn message_id(&mut self) -> Option<MessageId>

Returns the ID of the incoming message that is being handled, or None if not executing an incoming message.

pub fn with_message_is_bouncing( self, message_is_bouncing: impl Into<Option<bool>>, ) -> Self

Configures the message_is_bouncing flag to return during the test.

pub fn set_message_is_bouncing( &mut self, message_is_bouncing: impl Into<Option<bool>>, ) -> &mut Self

Configures the message_is_bouncing flag to return during the test.

pub fn message_is_bouncing(&mut self) -> Option<bool>

Returns true if the incoming message was rejected from the original destination and is now bouncing back, or None if not executing an incoming message.

pub fn with_authenticated_caller_id( self, authenticated_caller_id: impl Into<Option<ApplicationId>>, ) -> Self

Configures the authenticated caller ID to return during the test.

pub fn set_authenticated_caller_id( &mut self, authenticated_caller_id: impl Into<Option<ApplicationId>>, ) -> &mut Self

Configures the authenticated caller ID to return during the test.

pub fn authenticated_caller_id(&mut self) -> Option<ApplicationId>

Returns the authenticated caller ID, if the caller configured it and if the current context is executing a cross-application call.

pub fn with_system_time(self, timestamp: Timestamp) -> Self

Configures the system time to return during the test.

pub fn set_system_time(&mut self, timestamp: Timestamp) -> &mut Self

Configures the system time to return during the test.

pub fn system_time(&mut self) -> Timestamp

Retrieves the current system time, i.e. the timestamp of the block in which this is called.

pub fn with_chain_balance(self, chain_balance: Amount) -> Self

Configures the chain balance to return during the test.

pub fn set_chain_balance(&mut self, chain_balance: Amount) -> &mut Self

Configures the chain balance to return during the test.

pub fn chain_balance(&mut self) -> Amount

Returns the current chain balance.

pub fn with_owner_balances( self, owner_balances: impl IntoIterator<Item = (AccountOwner, Amount)>, ) -> Self

Configures the balances on the chain to use during the test.

pub fn set_owner_balances( &mut self, owner_balances: impl IntoIterator<Item = (AccountOwner, Amount)>, ) -> &mut Self

Configures the balances on the chain to use during the test.

pub fn with_owner_balance(self, owner: AccountOwner, balance: Amount) -> Self

Configures the balance of one account on the chain to use during the test.

pub fn set_owner_balance( &mut self, owner: AccountOwner, balance: Amount, ) -> &mut Self

Configures the balance of one account on the chain to use during the test.

pub fn owner_balance(&mut self, owner: AccountOwner) -> Amount

Returns the balance of one of the accounts on this chain.

pub fn send_message( &mut self, destination: impl Into<Destination>, message: Application::Message, )

Schedules a message to be sent to this application on another chain.

pub fn prepare_message( &mut self, message: Application::Message, ) -> MessageBuilder<Application::Message>

Returns a MessageBuilder to prepare a message to be sent.

pub fn created_send_message_requests( &self, ) -> MutexGuard<'_, Vec<SendMessageRequest<Application::Message>>>

Returns the list of SendMessageRequests created so far during the test.

pub fn subscribe(&mut self, chain: ChainId, channel: ChannelName)

Subscribes to a message channel from another chain.

pub fn subscribe_requests(&self) -> &[(ChainId, ChannelName)]

Returns the list of requests to subscribe to channels made in the test so far.

pub fn unsubscribe(&mut self, chain: ChainId, channel: ChannelName)

Unsubscribes to a message channel from another chain.

pub fn unsubscribe_requests(&self) -> &[(ChainId, ChannelName)]

Returns the list of requests to unsubscribe to channels made in the test so far.

pub fn transfer( &mut self, source: Option<AccountOwner>, destination: Account, amount: Amount, )

Transfers an amount of native tokens from source owner account (or the current chain’s balance) to destination.

pub fn outgoing_transfers(&self) -> &HashMap<Account, Amount>

Returns the outgoing transfers scheduled during the test so far.

pub fn claim(&mut self, source: Account, destination: Account, amount: Amount)

Claims an amount of native tokens from a source account to a destination account.

pub fn claim_requests(&self) -> &[ClaimRequest]

Returns the list of claims made during the test so far.

pub fn with_chain_ownership(self, chain_ownership: ChainOwnership) -> Self

Configures the chain ownership configuration to return during the test.

pub fn set_chain_ownership( &mut self, chain_ownership: ChainOwnership, ) -> &mut Self

Configures the chain ownership configuration to return during the test.

pub fn chain_ownership(&mut self) -> ChainOwnership

Retrieves the owner configuration for the current chain.

pub fn with_can_close_chain(self, can_close_chain: bool) -> Self

Configures if the application being tested is allowed to close the chain its in.

pub fn set_can_close_chain(&mut self, can_close_chain: bool) -> &mut Self

Configures if the application being tested is allowed to close the chain its in.

pub fn close_chain(&mut self) -> Result<(), CloseChainError>

Closes the current chain. Returns an error if the application doesn’t have permission to do so.

pub fn add_expected_open_chain_call( &mut self, ownership: ChainOwnership, application_permissions: ApplicationPermissions, balance: Amount, message_id: MessageId, )

Adds an expected call to open_chain, and the message ID that should be returned.

pub fn open_chain( &mut self, ownership: ChainOwnership, application_permissions: ApplicationPermissions, balance: Amount, ) -> (MessageId, ChainId)

Opens a new chain, configuring it with the provided chain_ownership, application_permissions and initial balance (debited from the current chain).

pub fn with_call_application_handler( self, handler: impl FnMut(bool, ApplicationId, Vec<u8>) -> Vec<u8> + 'static, ) -> Self

Configures the handler for cross-application calls made during the test.

pub fn set_call_application_handler( &mut self, handler: impl FnMut(bool, ApplicationId, Vec<u8>) -> Vec<u8> + 'static, ) -> &mut Self

Configures the handler for cross-application calls made during the test.

pub fn call_application<A: ContractAbi + Send>( &mut self, authenticated: bool, application: ApplicationId<A>, call: &A::Operation, ) -> A::Response

Calls another application.

pub fn emit(&mut self, name: StreamName, key: &[u8], value: &[u8])

Adds a new item to an event stream.

pub fn add_expected_service_query<A: ServiceAbi + Send>( &mut self, application_id: ApplicationId<A>, query: A::Query, response: A::QueryResponse, )

Adds an expected query_service call`, and the response it should return in the test.

pub fn add_expected_post_request( &mut self, url: String, payload: Vec<u8>, response: Vec<u8>, )

Adds an expected http_post call, and the response it should return in the test.

pub fn add_expected_read_data_blob_requests( &mut self, hash: DataBlobHash, response: Vec<u8>, )

Adds an expected read_data_blob call, and the response it should return in the test.

pub fn add_expected_assert_data_blob_exists_requests( &mut self, hash: DataBlobHash, response: Option<()>, )

Adds an expected assert_data_blob_exists call, and the response it should return in the test.

pub fn query_service<A: ServiceAbi + Send>( &mut self, application_id: ApplicationId<A>, query: A::Query, ) -> A::QueryResponse

Queries an application service as an oracle and returns the response.

Should only be used with queries where it is very likely that all validators will compute the same result, otherwise most block proposals will fail.

Cannot be used in fast blocks: A block using this call should be proposed by a regular owner, not a super owner.

pub fn http_post(&mut self, url: &str, payload: Vec<u8>) -> Vec<u8>

Makes a GET request to the given URL as an oracle and returns the JSON part, if any.

Should only be used with queries where it is very likely that all validators will receive the same response, otherwise most block proposals will fail.

Cannot be used in fast blocks: A block using this call should be proposed by a regular owner, not a super owner.

pub fn assert_before(&mut self, timestamp: Timestamp)

Panics if the current time at block validation is >= timestamp. Note that block validation happens at or after the block timestamp, but isn’t necessarily the same.

Cannot be used in fast blocks: A block using this call should be proposed by a regular owner, not a super owner.

pub fn read_data_blob(&mut self, hash: &DataBlobHash) -> Vec<u8>

Reads a data blob with the given hash from storage.

pub fn assert_data_blob_exists(&mut self, hash: DataBlobHash)

Asserts that a blob with the given hash exists in storage.