Module linera_sdk::linera_base_types

Types reexported from linera_base.

Structs

Account

A system account.

Amount

A non-negative amount of tokens.

ApplicationId

A unique identifier for a user application from a blob.

ApplicationPermissions

Permissions for applications on a chain.

Blob

A blob of binary data, with its hash.

BlobContent

A blob of binary data.

BlobId

A content-addressed blob ID i.e. the hash of the BlobContent.

BlockHeight

A block height to identify blocks in a chain.

Bytecode

A WebAssembly module’s bytecode.

ChainId

The unique identifier (UID) of a chain. This is currently computed as the hash value of a ChainDescription.

ChainOwnership

Represents the owner(s) of a chain.

ChannelFullName

A channel name together with its application ID.

ChannelName

The name of a subscription channel.

CompressedBytecode

A compressed WebAssembly module’s bytecode.

CryptoHash

A Keccak256 value.

CryptoHashVec

A vector of cryptographic hashes. This is used to represent a hash of a list of hashes.

Ed25519PublicKey

An Ed25519 signature public key.

Ed25519SecretKey

An Ed25519 secret key.

Ed25519Signature

An Ed25519 signature.

Event

An event recorded in an executed block.

EventId

An event identifier.

MessageId

The index of a message in a chain.

ModuleId

A unique identifier for a module.

Owner

The owner of a chain. This is currently the hash of the owner’s public key used to verify signatures.

Resources

Resources that an application may spend during the execution of transaction or an application call.

Secp256k1PublicKey

A secp256k1 public key.

Secp256k1SecretKey

A secp256k1 secret key.

Secp256k1Signature

A secp256k1 signature.

SendMessageRequest

A request to send a message.

StreamId

An event stream ID.

StreamName

The name of an event stream.

TestString

A BCS-signable struct for testing.

TimeDelta

A duration in microseconds.

TimeoutConfig

The timeout configuration: how long fast, multi-leader and single-leader rounds last.

Timestamp

A timestamp, in microseconds since the Unix epoch.

UserApplicationDescription

Description of the necessary information to run a user application used within blobs.

Enums

AccountOwner

An account owner.

AccountPublicKey

The public key of a chain owner. The corresponding private key is allowed to propose blocks on the chain and transfer account’s tokens.

AccountSecretKey

The private key of a chain owner.

AccountSignature

The signature of a chain owner.

ArithmeticError

An error type for arithmetic errors.

BcsHexParseError

An error trying to parse the hex-digits of a BCS-encoded value.

BlobType

The type of the blob. Should be a 1:1 mapping of the types in Blob.

ChainDescription

How to create a chain.

ChangeApplicationPermissionsError

Errors that can happen when attempting to change the application permissions.

CloseChainError

Errors that can happen when attempting to close a chain.

CryptoError

Error type for cryptographic errors.

DecompressionError

A type for errors happening during decompression.

Destination

The destination of a message, relative to a particular application.

GenericApplicationId

A unique identifier for an application.

OracleResponse

A record of a single oracle response.

ParseAmountError

Round

An identifier for successive attempts to decide a value in a consensus protocol.

SignatureScheme

Signature scheme used for the public key.

Traits

Abi

A trait that includes all the types exported by a Linera application (both contract and service).

BcsHashable

Activate the blanket implementation of Hashable based on serde and BCS.

BcsSignable

Activate the blanket implementation of Signable based on serde and BCS.

ContractAbi

A trait that includes all the types exported by a Linera application contract.

CryptoRng

Wrapper around rand::CryptoRng and rand::RngCore.

HasTypeName

Something that we know how to hash and sign.

Hashable

Something that we know how to hash.

ServiceAbi

A trait that includes all the types exported by a Linera application service.

WithContractAbi

Marker trait to help importing contract types.

WithServiceAbi

Marker trait to help importing service types.

Type Aliases

UserApplicationId

Alias for ApplicationId. Use this alias in the core protocol where the distinction with the more general enum GenericApplicationId matters.

ValidatorKeypair

The key pair of a validator.

ValidatorPublicKey

The public key of a validator.

ValidatorSecretKey

The private key of a validator.

ValidatorSignature

The signature of a validator.