Module linera_sdk::test

Helper types for writing integration tests for WebAssembly applications.

Integration tests are usually written in the tests directory in the root of the crate’s directory (i.e., beside the src directory). Linera application integration tests should be executed targeting the host architecture, instead of targeting wasm32-unknown-unknown like done for unit tests.

Structs

ActiveChain

A reference to a single microchain inside a TestValidator.

BlockBuilder

A helper type to build Blocks using the builder pattern, and then signing them into Certificates using a TestValidator.

TestValidator

A minimal validator implementation suited for tests.