Trait linera_sdk::Contract

pub trait Contract: Sized {
    type Error: Error;
    type Storage;

    fn initialize<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        argument: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn execute_operation<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        operation: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn execute_effect<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 EffectContext,
        effect: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn handle_application_call<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        argument: &'life2 [u8],
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn handle_session_call<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        session: Session,
        argument: &'life2 [u8],
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<SessionCallResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

The public entry points provided by a contract.

Required Associated Types

type Error: Error

Message reports for application execution errors.

type Storage

Tag the contract with the desired state management runtime.

Required Methods

fn initialize<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 OperationContext,
argument: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Initialize the application on the chain that created it.

fn execute_operation<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 OperationContext,
operation: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Apply an operation from the current block.

fn execute_effect<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 EffectContext,
effect: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Apply an effect originating from a cross-chain message.

fn handle_application_call<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 CalleeContext,
argument: &'life2 [u8],
forwarded_sessions: Vec<SessionId>
) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Allow an operation or an effect of other applications to call into this application.

fn handle_session_call<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 CalleeContext,
session: Session,
argument: &'life2 [u8],
forwarded_sessions: Vec<SessionId>
) -> Pin<Box<dyn Future<Output = Result<SessionCallResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Allow an operation or an effect of other applications to call into a session that we previously created.