Trait linera_sdk::Contract

pub trait Contract: Sized {
    type Error: Error;
    type Storage;

    fn initialize<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        argument: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn execute_operation<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        operation: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn execute_effect<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 EffectContext,
        effect: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn handle_application_call<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        argument: &'life2 [u8],
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn handle_session_call<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        session: Session,
        argument: &'life2 [u8],
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<SessionCallResult, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

The public entry points provided by an application’s contract.

Execution of these endpoints consume fuel, because they can change the application’s state and are therefore consensus criticial.

Required Associated Types

type Error: Error

Message reports for application execution errors.

type Storage

The desired storage backend to use to store the application’s state.

Required Methods

fn initialize<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 OperationContext,
argument: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Initializes the application on the chain that created it.

This is only called once when the application is created and only on the microchain that created the application.

Deployment on other microchains will use the Default implementation of the application state if SimpleStateStorage is used, or the Default value of all sub-views in the state if the ViewStateStorage is used.

Returns an ExecutionResult, which can contain subscription or unsubscription requests to channels and effects to be sent to this application on another microchain.

fn execute_operation<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 OperationContext,
operation: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Applies an operation from the current block.

Operations are created by users and added to blocks, serving as the starting point for an application’s execution.

Returns an ExecutionResult, which can contain subscription or unsubscription requests to channels and effects to be sent to this application on another microchain.

fn execute_effect<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 EffectContext,
effect: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = Result<ExecutionResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Applies an effect originating from a cross-chain message.

Effects are messages sent across microchains. These messages are created and received by the same application. Effects can be either single-sender and single-receiver, or single-sender and multiple-receivers. The former allows sending cross-chain messages to the application on some other specific microchain, while the latter uses broadcast channels to send a message to multiple other microchains where the application is subscribed to a sender channel on this microchain.

For an effect to be executed, a user must mark it to be received in a block of the receiver microchain.

Returns an ExecutionResult, which can contain effects to be sent to this application on another microchain and subscription or unsubscription requests to channels.

fn handle_application_call<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 CalleeContext,
argument: &'life2 [u8],
forwarded_sessions: Vec<SessionId>
) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Handles a call from another application.

Cross-application calls are a way for applications to interact inside a microchain. An application can call any other application available on the microchain the execution is taking place on.

Use the Self::call_application method generated by the contract! macro to call another application.

Returns an ApplicationCallResult, which contains:

a response value sent to the caller application;

a list of newly created Sessions sent to the caller application;

an ExecutionResult with effects to be sent to this application on other microchains and channel subscription and unsubscription requests.

See Self::handle_session_call for more information on

fn handle_session_call<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
context: &'life1 CalleeContext,
session: Session,
argument: &'life2 [u8],
forwarded_sessions: Vec<SessionId>
) -> Pin<Box<dyn Future<Output = Result<SessionCallResult, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Handles a call into a Session created by this application.

Sessions are another way for applications to interact inside a microchain. Sessions are very similar to cross-application calls (see Self::handle_application_call), but each one of them has a separate state, that’s also separate from the application state.

This allows moving parts of the application’s state out into sessions, which are sent to other applications. These sessions can be freely moved between other applications, and the application that created the session does not have to keep track of which application is the current owner of the session.

Other applications receive SessionIds to reference the sessions, so that they can’t read or alter the session’s state. The only way for an application to interact with another application’s session is through calls into that session.

At the end of the execution of a block, no sessions should be alive. If a block’s execution ends with any leaked sessions, the block is rejected. This means that all sessions should be called at least once in a way that the call terminates the session (by returning None as the updated session data in SessionCallResult::data).

Use the Self::call_application method generated by the contract! macro to call another application.

Returns a SessionCallResult after the call into the session has finished executing, which contains:

the updated session data, or None if the session should be terminated;

an ApplicationCallResult, which contains:

a response value sent to the caller application;

a list of newly created Sessions sent to the caller application;

an ExecutionResult with effects to be sent to this application on other microchains and channel subscription and unsubscription requests.