Trait linera_sdk::Contract

pub trait Contract: WithContractAbi + ContractAbi + Send + Sized {
    type Error: Error + From<Error> + From<Error> + 'static;
    type Storage: ContractStateStorage<Self> + Send + 'static;

    // Required methods
    fn initialize<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        argument: Self::InitializationArgument
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn execute_operation<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        context: &'life1 OperationContext,
        operation: Self::Operation
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn execute_message<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        context: &'life1 MessageContext,
        message: Self::Message
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn handle_application_call<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        argument: Self::ApplicationCall,
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult<Self::Message, Self::Response, Self::SessionState>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn handle_session_call<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        context: &'life1 CalleeContext,
        session: Self::SessionState,
        argument: Self::SessionCall,
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<SessionCallResult<Self::Message, Self::Response, Self::SessionState>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn call_application<'life0, 'life1, 'async_trait, A>(
        &'life0 mut self,
        authenticated: bool,
        application: ApplicationId<A>,
        call: &'life1 A::ApplicationCall,
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<(A::Response, Vec<SessionId>), Self::Error>> + Send + 'async_trait>>
       where A: 'async_trait + ContractAbi + Send,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn call_session<'life0, 'life1, 'async_trait, A>(
        &'life0 mut self,
        authenticated: bool,
        session: SessionId<A>,
        call: &'life1 A::SessionCall,
        forwarded_sessions: Vec<SessionId>
    ) -> Pin<Box<dyn Future<Output = Result<(A::Response, Vec<SessionId>), Self::Error>> + Send + 'async_trait>>
       where A: 'async_trait + ContractAbi + Send,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn parameters() -> Result<Self::Parameters, Self::Error> { ... }
    fn current_application_id() -> ApplicationId<Self::Abi> { ... }
}

The contract interface of a Linera application.

As opposed to the Service interface of an application, contract entry points are triggered by the execution of blocks in a chain. Their execution may modify storage and is gas-metered.

Below we use the word “transaction” to refer to the current operation or message being executed.

Required Associated Types

type Error: Error + From<Error> + From<Error> + 'static

The type used to report errors to the execution environment.

Errors are not recoverable and always interrupt the current transaction. To return recoverable errors in the case of application calls and session calls, you may use the response types.

type Storage: ContractStateStorage<Self> + Send + 'static

The desired storage backend used to store the application’s state.

Currently, the two supported backends are SimpleStateStorage or ViewStateStorage. Accordingly, this associated type may be defined as type Storage = SimpleStateStorage<Self> or type Storage = ViewStateStorage<Self>.

The first deployment on other chains will use the Default implementation of the application state if SimpleStateStorage is used, or the Default value of all sub-views in the state if the ViewStateStorage is used.

Required Methods

fn initialize<'life0, 'life1, 'async_trait>( &'life0 mut self, context: &'life1 OperationContext, argument: Self::InitializationArgument ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Initializes the application on the chain that created it.

This is only called once when the application is created and only on the microchain that created the application.

Returns an ExecutionResult, which can contain subscription or unsubscription requests to channels and messages to be sent to this application on another chain.

fn execute_operation<'life0, 'life1, 'async_trait>( &'life0 mut self, context: &'life1 OperationContext, operation: Self::Operation ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Applies an operation from the current block.

Operations are created by users and added to blocks, serving as the starting point for an application’s execution.

Returns an ExecutionResult, which can contain subscription or unsubscription requests to channels and messages to be sent to this application on another chain.

fn execute_message<'life0, 'life1, 'async_trait>( &'life0 mut self, context: &'life1 MessageContext, message: Self::Message ) -> Pin<Box<dyn Future<Output = Result<ExecutionResult<Self::Message>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Applies a message originating from a cross-chain message.

Messages are messages sent across chains. These messages are created and received by the same application. Messages can be either single-sender and single-receiver, or single-sender and multiple-receivers. The former allows sending cross-chain messages to the application on some other specific chain, while the latter uses broadcast channels to send a message to multiple other chains where the application is subscribed to a sender channel on this chain.

For a message to be executed, a user must mark it to be received in a block of the receiver chain.

Returns an ExecutionResult, which can contain messages to be sent to this application on another chain and subscription or unsubscription requests to channels.

fn handle_application_call<'life0, 'life1, 'async_trait>( &'life0 mut self, context: &'life1 CalleeContext, argument: Self::ApplicationCall, forwarded_sessions: Vec<SessionId> ) -> Pin<Box<dyn Future<Output = Result<ApplicationCallResult<Self::Message, Self::Response, Self::SessionState>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handles a call from another application.

Cross-application calls allow applications to interact inside a chain. An application can call any other application available on the chain the execution is taking place on.

Use the Self::call_application method generated by the contract! macro to call another application.

Returns an ApplicationCallResult, which contains:

a return value sent to the caller application;

a list of new session states; the newly-created sessions will be owned by the caller application;

an ExecutionResult with messages to be sent to this application on other chains and channel subscription and unsubscription requests.

See Self::handle_session_call for more information on

fn handle_session_call<'life0, 'life1, 'async_trait>( &'life0 mut self, context: &'life1 CalleeContext, session: Self::SessionState, argument: Self::SessionCall, forwarded_sessions: Vec<SessionId> ) -> Pin<Box<dyn Future<Output = Result<SessionCallResult<Self::Message, Self::Response, Self::SessionState>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handles a call into a session created by this application.

Sessions are another way for applications to interact inside a chain. Sessions are very similar to cross-application calls (see Self::handle_application_call), but each one of them has a separate state, in addition to the application state.

Sessions allow representing transferrable objects (typically assets and liabilities) for the duration of a transaction. A session is initially owned by the application that made a call to create it. Ownership over a session can be transferred to other applications by passing them as session arguments. The execution environment keeps track of which application is the current owner of the session.

The state of a session is only visible to the application that created it (and which handles the session calls). Other applications only see the SessionId and may only interact with the session through session calls.

At the end of the execution of a block, no sessions should be alive. If a block’s execution ends with any leaked sessions, the block is rejected. This means that all sessions should be called at least once in a way that the call terminates the session (by returning None as the updated session data in SessionCallResult::new_state).

Use the Self::call_application method generated by the contract! macro to call another application.

Returns a SessionCallResult after the call into the session has finished executing, which contains:

the updated session state, or None if the session should be terminated;

an ApplicationCallResult, which contains:

a return value sent to the caller application;

a list of new session states; the newly-created sessions will be owned by the caller application;

an ExecutionResult with messages to be sent to this application on other chains and channel subscription and unsubscription requests.

Provided Methods

fn call_application<'life0, 'life1, 'async_trait, A>( &'life0 mut self, authenticated: bool, application: ApplicationId<A>, call: &'life1 A::ApplicationCall, forwarded_sessions: Vec<SessionId> ) -> Pin<Box<dyn Future<Output = Result<(A::Response, Vec<SessionId>), Self::Error>> + Send + 'async_trait>>where A: 'async_trait + ContractAbi + Send, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Calls another application.

fn call_session<'life0, 'life1, 'async_trait, A>( &'life0 mut self, authenticated: bool, session: SessionId<A>, call: &'life1 A::SessionCall, forwarded_sessions: Vec<SessionId> ) -> Pin<Box<dyn Future<Output = Result<(A::Response, Vec<SessionId>), Self::Error>> + Send + 'async_trait>>where A: 'async_trait + ContractAbi + Send, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Calls a session from another application.

fn parameters() -> Result<Self::Parameters, Self::Error>

Retrieves the parameters of the application.

fn current_application_id() -> ApplicationId<Self::Abi>

Retrieves the current application ID.