Trait linera_sdk::Service

pub trait Service: ServiceAbi {
    type Error: Error + From<Error>;
    type Storage: ServiceStateStorage;

    // Required method
    fn query_application<'life0, 'async_trait>(
        self: Arc<Self>,
        context: &'life0 QueryContext,
        argument: Self::Query
    ) -> Pin<Box<dyn Future<Output = Result<Self::QueryResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn parameters() -> Result<Self::Parameters, Self::Error> { ... }
}

The service interface of a Linera application.

As opposed to the Contract interface of an application, service entry points are triggered by JSON queries (typically GraphQL). Their execution cannot modify storage and is not gas-metered.

Required Associated Types

type Error: Error + From<Error>

Type used to report errors to the execution environment.

Errors are not recoverable and always interrupt the current query.

type Storage: ServiceStateStorage

The desired storage backend used to store the application’s state.

Currently, the two supported backends are SimpleStateStorage or ViewStateStorage. Accordingly, this associated type may be defined as type Storage = SimpleStateStorage<Self> or type Storage = ViewStateStorage<Self>.

Required Methods

fn query_application<'life0, 'async_trait>( self: Arc<Self>, context: &'life0 QueryContext, argument: Self::Query ) -> Pin<Box<dyn Future<Output = Result<Self::QueryResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Executes a read-only query on the state of this application.

Provided Methods

fn parameters() -> Result<Self::Parameters, Self::Error>

Retrieves the parameters of the application.