linera_sdk::util

Trait BlockingWait

pub trait BlockingWait {
    type Output;

    // Required method
    fn blocking_wait(self) -> Self::Output;
}

An extension trait to block on a Future until it completes.

Required Associated Types

type Output

The type returned by the Future.

Required Methods

fn blocking_wait(self) -> Self::Output

Waits for the Future to complete in a blocking manner.

Effectively polls the Future repeatedly until it returns Poll::Ready.