Module linera_sdk::views

Helper types for using linera_views to store application state.

Re-exports

pub use linera_views;

Enums

ViewError

Main error type for the crate.

Traits

CustomSerialize

Sometimes we need a serialization that is different from the usual one and for example preserves order. The {to/from}_custom_bytes has to be coherent with the Borrow trait.

RootView

A View whose staged modifications can be saved in storage.

View

A view gives exclusive access to read and write the data stored at an underlying address in storage.

Type Aliases

ByteCollectionView

An alias to linera_views::collection_view::ByteCollectionView that uses the WebAssembly specific ViewStorageContext.

ByteMapView

An alias to linera_views::map_view::ByteMapView that uses the WebAssembly specific ViewStorageContext.

ByteSetView

An alias to linera_views::set_view::ByteSetView that uses the WebAssembly specific ViewStorageContext.

CollectionView

An alias to linera_views::collection_view::CollectionView that uses the WebAssembly specific ViewStorageContext.

CustomCollectionView

An alias to linera_views::collection_view::CustomCollectionView that uses the WebAssembly specific ViewStorageContext.

CustomMapView

An alias to linera_views::map_view::CustomMapView that uses the WebAssembly specific ViewStorageContext.

CustomSetView

An alias to linera_views::set_view::CustomSetView that uses the WebAssembly specific ViewStorageContext.

LogView

An alias to linera_views::log_view::LogView that uses the WebAssembly specific ViewStorageContext.

MapView

An alias to linera_views::map_view::MapView that uses the WebAssembly specific ViewStorageContext.

QueueView

An alias to linera_views::queue_view::QueueView that uses the WebAssembly specific ViewStorageContext.

ReadGuardedView

An alias to linera_views::collection_view::ReadGuardedView that uses the WebAssembly specific ViewStorageContext.

RegisterView

An alias to linera_views::register_view::RegisterView that uses the WebAssembly specific ViewStorageContext.

SetView

An alias to linera_views::set_view::SetView that uses the WebAssembly specific ViewStorageContext.

ViewStorageContext

Implementation of linera_views::common::Context to be used for data storage by Linera applications.

Derive Macros

RootView

View