Module linera_sdk::views

Helper types for using linera_views to store application state.

Modules

batch

The definition of the batches for writing in the database. A set of functionalities for building batches to be written into the database. A batch can contain three kinds of operations on a key/value store:

collection_view

The CollectionView implements a map structure whose keys are ordered and the values are views.

common

The definitions used for the memory, RocksDB, and DynamoDB backends. This provides several functionalities for the handling of data. The most important traits are:

crypto

dynamo_db

Helper definitions for DynamoDB storage.

futures

Abstractions for asynchronous programming.

generic_array

This crate implements a structure that can be used as a generic array type. Core Rust array types [T; N] can’t be used generically with respect to N, so for example this:

hashable_wrapper

Wrapping a view to compute a hash.

key_value_store_view

The implementation of a key-value store view.

localstack

Helper types for interfacing with a LocalStack instance.

log_view

The LogView implements a log list that can be pushed.

lru_caching

The LRU (least recently used) caching.

map_view

The MapView implements a map with ordered keys. The MapView implements a map that can be modified.

memory

Helper definitions for in-memory storage.

queue_view

The QueueView implements a queue that can push on the back and delete on the front.

reentrant_collection_view

The ReentrantCollectionView implements a map structure whose keys are ordered and the values are views with concurrent access.

register_view

The RegisterView implements a register for a single value.

rocksdb

Helper definitions for Rocksdb storage.

serde

Serde

set_view

The SetView implements a set with ordered entries.

sha3

An implementation of the SHA-3 cryptographic hash algorithms.

test_utils

Helper types for tests.

views

The definition of the View and related traits.

Constants

LOAD_VIEW_COUNTER

The metric counting how often a view is read from storage.

SAVE_VIEW_COUNTER

The metric counting how often a view is written from storage.

Functions

increment_counter

Increments the metrics counter with the given name, with the struct and base key as labels.

Type Definitions

ByteCollectionView

An alias to collection_view::ByteCollectionView that uses the WebAssembly specific ViewStorageContext.

ByteMapView

An alias to map_view::ByteMapView that uses the WebAssembly specific ViewStorageContext.

ByteSetView

An alias to set_view::ByteSetView that uses the WebAssembly specific ViewStorageContext.

CollectionView

An alias to collection_view::CollectionView that uses the WebAssembly specific ViewStorageContext.

CustomCollectionView

An alias to collection_view::CustomCollectionView that uses the WebAssembly specific ViewStorageContext.

CustomMapView

An alias to map_view::CustomMapView that uses the WebAssembly specific ViewStorageContext.

CustomSetView

An alias to set_view::CustomSetView that uses the WebAssembly specific ViewStorageContext.

LogView

An alias to log_view::LogView that uses the WebAssembly specific ViewStorageContext.

MapView

An alias to map_view::MapView that uses the WebAssembly specific ViewStorageContext.

QueueView

An alias to queue_view::QueueView that uses the WebAssembly specific ViewStorageContext.

ReadGuardedView

An alias to collection_view::ReadGuardedView that uses the WebAssembly specific ViewStorageContext.

RegisterView

An alias to register_view::RegisterView that uses the WebAssembly specific ViewStorageContext.

SetView

An alias to set_view::SetView that uses the WebAssembly specific ViewStorageContext.

ViewStorageContext

Implementation of linera_views::common::Context to be used for data storage by Linera applications.

Attribute Macros

async_trait