Trait linera_sdk::views::CustomSerialize

pub trait CustomSerialize: Sized {
    // Required methods
    fn to_custom_bytes(&self) -> Result<Vec<u8>, ViewError>;
    fn from_custom_bytes(short_key: &[u8]) -> Result<Self, ViewError>;
}

Sometimes we need a serialization that is different from the usual one and for example preserves order. The {to/from}_custom_bytes has to be coherent with the Borrow trait.

Required Methods

fn to_custom_bytes(&self) -> Result<Vec<u8>, ViewError>

Serializes the value

fn from_custom_bytes(short_key: &[u8]) -> Result<Self, ViewError>

Deserialize the vector

Object Safety

This trait is not object safe.